chore(trait): cleaned up/moved class traits to proper folders [#5]
