+++ src/Entities/Message/DocumentEntity.php
@@ -4,7 +4,7 @@ namespace The42dx\Whatsapp\Entities\Message;
 
 use The42dx\Whatsapp\Abstracts\MediaEntity;
 use The42dx\Whatsapp\Contracts\Entity;
-use The42dx\Whatsapp\Traits\HasCaption;
+use The42dx\Whatsapp\Entities\Traits\HasCaption;
 
 /**
  * DocumentEntity
